Add PGE header and create a game from template.
We copy in the olcPixelGameEngine.h file and add it to our solution.
We also add a blank "main.cpp" and populate it with the contents of
the template available in the olcPixelGameEngine.h header, taking care
to rename our game class to match our needs.

Running our project will render a default PixelGameEngine scene: a
256x240 canvas of random pixels, magnified 4x.