Add constants and a way to "Sample" single pixels.
We define a few constants for window geometry and begin implementing our
rendering process by looping over the rows and columns of the scene and
calling a "Sample" function that takes a floating-point x,y position on
the viewport and returns a olc::Pixel for that location.

Running our project will now render a 250x250 canvas at 2x
magnification. Our magenta fill has been replaced with a color pattern
converging in the center of the canvas.