Add fog color and a way to sample rays.
To prevent our scene from extending into infinity, and to have something
to show when a ray doesn't hit *anything*, we add a new constant: a
"fog" color.
Additionally, we add a more specific function, SampleRay, that is called
by Sample to return the color (or absence thereof) of a ray as it
extends into our scene. For now, still, this returns a color relative to
the X and Y coordinate in our scene.

Running our project produces no differences from our last commit.