Implement ray-Sphere intersection.
We'll need to overload some operators for a vf3d: subtraction, and
dot-product. A dot-product is a useful way of comparing two vectors to
determine if they are similar.

We'll also implement the equation for an intersection between a ray and
a Sphere. I'm not going to go into depth explaining the geometry here:
this is a well-documented process and can be researched separately.

Running our project will now render a (highly aliased and
flatly-colored) Sphere!