Add perspective rendering and depth sorting.
First we'll add some additional Spheres to our scene at different
Z-depths. If we run our project now, you'll see that the Spheres added
to our scene last are drawn in front of the earlier ones, even if they
are further away.

To remedy that, we up date our hit check in SampleRay to select the
object whose intersection is nearest to the ray origin. Now if we run
our project, the Spheres are properly sorted. However, you'll notice
that all three Spheres are the same size, despite being different
distances from the camera.

To fix this, we'll need to add perspective to our camera. We'll do this
in a very simplistic manner, by having all of the rays originate from
some point, and pointing towards what you can think of as a "virtual
canvas" some distance in front of that origin point. By normalizing this
ray we get rays properly fanning out in a perspective.

Running our project will now produce a proper perspective rendering of
our three flat-shaded Spheres, at the correct depths.